+++ Example/Vanilla SwiftUI Example/Vanilla SwiftUI Example/Train.swift
@@ -2,5 +2,5 @@ import Foundation
 
 struct Train: Hashable {
   let name: String
-  let capacity: Int64
+  let capacity: Int
 }
